Refactored scenes to allow scene dependencies, and updated scenes to work as multi-scenes with scene dependencies additively loaded.

+++ Assets/Code/Runtime/Characters/IHasHealth.cs
@@ -2,7 +2,7 @@
 {
     public interface IHasHealth
     {
-        public void TakeDamage(float damage);
-        public float GetCurrentHealth();
+        public void TakeDamage(int damage);
+        public int GetCurrentHealth();
     }
 }
